Place garbage collector calls after user code
The note [Heap checks] talks about this layout for heap checks:

        f( f_clo, x:ptr, y:ptr, ...etc... ) {
          T: Hp = Hp+8
             if Hp > HpLim goto L
             ...
          L: HpAlloc = 8
             call gc()  -- Needs an info table
             goto T }

But the code generator generates something like:

        f( f_clo, x:ptr, y:ptr, ...etc... ) {
          T: Hp = Hp+8
             if Hp > HpLim goto L
             HpAlloc = 8
             call gc()  -- Needs an info table
             goto T
          L:  ...

In cases the heap check doesn't trigger garbage collection the jump is
taken. This seems to be the majority of the cases. This patch changes
the heap check according to the note. This allows for more tighter entry
code and hopefully some (minor) performance gains.

Unfortunately I cannot see why the code diverged from the note as
git blame shows roughly the same commit for both note and code.