Abstract over the way eventlogs are flushed
Summary:
Currently eventlog data is always written to a file
`progname.eventlog`. This patch introduces the `flushEventLog` field
in `RtsConfig` which allows to customize the writing of eventlog data.

One possible scenario is the ongoing live-profile-monitor effort by
@NCrashed which slurps all eventlog data through `fluchEventLog`.

`flushEventLog` takes a buffer with eventlog data and its size and
returns `false` (0) in case eventlog data could not be procesed.

Reviewers: austin, bgamari, erikd, simonmar

Subscribers: thomie, NCrashed

Differential Revision: https://phabricator.haskell.org/D2934