Introduce StgFreeVar
A new notion for free variable lists in STG which can express
indirections in free variable lists. This allows us to implement
sharing for nested closures.

```
let [f{a b c}] = \[] -> ...
```

Only the free variable `f` is placed into the closure. `a`, `b` and
`c` are introduced through an indirection via `f`.